+++ Events/CarrierJumpEngagedEvent.cs
@@ -49,7 +49,7 @@ namespace EddiEvents
 
         // These properties are not intended to be user facing
 
-        public long? carrierId { get; private set; }
+        public long? carrierID { get; private set; }
 
         public CarrierJumpEngagedEvent(DateTime timestamp, string systemName, ulong systemAddress, string originSystemName, ulong originSystemAddress, string bodyName, long? bodyId, long? carrierId) : base(timestamp, NAME)
         {
@@ -64,7 +64,7 @@ namespace EddiEvents
             this.bodyId = bodyId;
 
             // Carrier
-            this.carrierId = carrierId;
+            this.carrierID = carrierId;
         }
     }
 }